Phase 5: Extract domain-specific hooks from App.jsx
Extract four domain-specific custom hooks from App.jsx, reducing it from
7,238 to 6,466 lines (-772 lines, -10.7%). Each hook owns its state
declarations, functions, and effects while App.jsx retains orchestration.

New hooks created (1,107 lines total):

- useFTCOfflineMode (145 lines): FTC local server probing and offline
  mode management. Contains 7 state vars (FTCOfflineAvailable,
  useFTCOffline, FTCKey, FTCServerURL, manualOfflineMode, ftcLeagues,
  ftcTypes), 4 functions (getFTCOfflineStatus, getFTCLeagues,
  requestFTCKey, checkFTCKey), and 2 effects (server status check on
  URL/mode change, manualOfflineMode reset on mode switch).

- useHighScores (578 lines): World stats, event high scores, and
  district/region/league high score management. Contains 5 state vars,
  7 functions (getWorldStats, getEventStats, getFrcDistrictHighScores,
  getFTCRegionHighScores, getFTCLeagueHighScores, normalizeFTCHighScores,
  calculateEventHighScores), and 2 effects (clear on ftcMode change,
  recalculate on schedule update).

- useCheesyArena (47 lines): Cheesy Arena server status probing.
  Contains 3 state vars (cheesyArenaAvailable, useCheesyArena setting,
  cheesyTeamList) and 1 probe-only function (getCheesyStatus).
  Orchestration (calling getTeamList/getSchedule) removed from probe
  since the loadSchedule parameter was unused at all call sites.

- useCommunityUpdates (337 lines): Community/scouting update fetching.
  Contains 2 state vars (communityUpdates, loadingCommunityUpdates),
  1 large function (getCommunityUpdates ~260 lines), and 1 effect
  (trigger on teamList changes).

Additional cleanup:

- Consolidated duplicate community update triggers: removed 2 redundant
  imperative getCommunityUpdates() calls from getTeamList() (OffSeason
  and FTC Offline branches), keeping only the OFFLINE branch call which
  is not covered by the useEffect. The useEffect on teamList changes is
  now the primary trigger for community updates.

- Moved getEffectiveTeamNumber function earlier in App.jsx (before hook
  calls) since it is passed as a parameter to useCommunityUpdates and
  also used by putTeamData.

All hooks follow the established pattern: receive inputs as an object
parameter, own their state/effects, and return all values and functions
for App.jsx to destructure and wire into contexts/orchestrators.
